refactored checkpoint form to use redux-forms to manage state
